Hooks can no longer be reeled in without being attatched, and regen.
The player regens once every 5 seconds, it does not currently scale with deaths.
Boosted the speed of hooks and how much they reel you in.
Increased the size of target hooks so they are a little more lenient to hit while moving.

The movement needs more work, it felt better when you could constantly reel.